+++ lib/src/ffi/native_module.dart
@@ -498,20 +498,8 @@ class NativeModule implements ffi.Finalizable {
 
     return TensorData(
       shape: shape,
-      dataType: _etDTypeToTensorType(dtype),
+      dataType: TensorType.fromExecuTorchValue(dtype.value),
       data: data,
     );
   }
-
-  /// Convert ETDType to TensorType.
-  static TensorType _etDTypeToTensorType(ETDType dtype) => switch (dtype) {
-        ETDType.ET_DTYPE_FLOAT32 => TensorType.float32,
-        ETDType.ET_DTYPE_FLOAT64 => TensorType.float32, // Fallback
-        ETDType.ET_DTYPE_INT64 => TensorType.int32, // Fallback
-        ETDType.ET_DTYPE_INT32 => TensorType.int32,
-        ETDType.ET_DTYPE_INT16 => TensorType.int32, // Fallback
-        ETDType.ET_DTYPE_INT8 => TensorType.int8,
-        ETDType.ET_DTYPE_UINT8 => TensorType.uint8,
-        ETDType.ET_DTYPE_BOOL => TensorType.uint8, // Fallback
-      };
 }